+++ src/pydocstyle/config.py
@@ -7,10 +7,7 @@ from collections import Set, namedtuple
 from re import compile as re
 
 
-try:  # Python 3.x
-    from ConfigParser import RawConfigParser
-except ImportError:  # Python 2.x
-    from configparser import RawConfigParser
+from configparser import RawConfigParser
 
 
 from .utils import __version__, log
@@ -129,7 +126,7 @@ class ConfigurationParser(object):
     def get_files_to_check(self):
         """Generate files and error codes to check on each one.
 
-        Walk dir trees under `self._arguments` and generate yield filnames
+        Walk dir trees under `self._arguments` and yield file names
         that `match` under each directory that `match_dir`.
         The method locates the configuration for each file name and yields a
         tuple of (filename, [error_codes]).
@@ -155,7 +152,7 @@ class ConfigurationParser(object):
         for name in self._arguments:
             if os.path.isdir(name):
                 for root, dirs, filenames in os.walk(name):
-                    config = self._get_config(root)
+                    config = self._get_config(os.path.abspath(root))
                     match, match_dir = _get_matches(config)
                     ignore_decorators = _get_ignore_decorators(config)
 
@@ -168,7 +165,7 @@ class ConfigurationParser(object):
                             yield (full_path, list(config.checked_codes),
                                    ignore_decorators)
             else:
-                config = self._get_config(name)
+                config = self._get_config(os.path.abspath(name))
                 match, _ = _get_matches(config)
                 ignore_decorators = _get_ignore_decorators(config)
                 if match(name):
@@ -256,10 +253,17 @@ class ConfigurationParser(object):
                 raise IllegalConfiguration('Configuration file {!r} specified '
                                            'via --config was not found.'
                                            .format(self._run_conf.config))
+
             if None in self._cache:
                 return self._cache[None]
             options, _ = self._read_configuration_file(self._run_conf.config)
-            config = self._create_check_config(options)
+
+            if options is None:
+                log.warning('Configuration file does not contain a '
+                            'pydocstyle section. Using default configuration.')
+                config = self._create_check_config(self._options)
+            else:
+                config = self._create_check_config(options)
 
         # Make the CLI always win
         final_config = {}
@@ -294,13 +298,17 @@ class ConfigurationParser(object):
         Returns (options, should_inherit).
 
         """
-        parser = RawConfigParser()
+        parser = RawConfigParser(inline_comment_prefixes=('#', ';'))
         options = None
         should_inherit = True
 
         if parser.read(path) and self._get_section_name(parser):
+            all_options = self._parser.option_list[:]
+            for group in self._parser.option_groups:
+                all_options.extend(group.option_list)
+
             option_list = dict([(o.dest, o.type or o.action)
-                                for o in self._parser.option_list])
+                                for o in all_options])
 
             # First, read the default values
             new_options, _ = self._parse_args([])
@@ -446,12 +454,18 @@ class ConfigurationParser(object):
 
         try:
             for part in code_parts:
-                if len(part) < 4:
-                    for code in codes:
-                        if code.startswith(part):
-                            expanded_codes.add(code)
-                else:
-                    expanded_codes.add(part)
+                # Dealing with split-lined configurations; The part might begin
+                # with a whitespace due to the newline character.
+                part = part.strip()
+                if not part:
+                    continue
+
+                codes_to_add = {code for code in codes
+                                if code.startswith(part)}
+                if not codes_to_add:
+                    log.warn('Error code passed is not a prefix of any known '
+                             'errors: %s', part)
+                expanded_codes.update(codes_to_add)
         except TypeError as e:
             raise IllegalConfiguration(e)
 
@@ -496,8 +510,8 @@ class ConfigurationParser(object):
         return any([getattr(options, opt) is not None for opt in
                     cls.BASE_ERROR_SELECTION_OPTIONS])
 
-    @staticmethod
-    def _fix_set_options(options):
+    @classmethod
+    def _fix_set_options(cls, options):
         """Alter the set options from None/strings to sets in place."""
         optional_set_options = ('ignore', 'select')
         mandatory_set_options = ('add_ignore', 'add_select')
@@ -506,9 +520,11 @@ class ConfigurationParser(object):
             """Split `value_str` by the delimiter `,` and return a set.
 
             Removes any occurrences of '' in the set.
+            Also expand error code prefixes, to avoid doing this for every
+            file.
 
             """
-            return set(value_str.split(',')) - {''}
+            return cls._expand_error_codes(set(value_str.split(',')) - {''})
 
         for opt in optional_set_options:
             value = getattr(options, opt)
@@ -530,7 +546,7 @@ class ConfigurationParser(object):
     @classmethod
     def _create_option_parser(cls):
         """Return an option parser to parse the command line arguments."""
-        from optparse import OptionParser
+        from optparse import OptionParser, OptionGroup
 
         parser = OptionParser(
             version=__version__,
@@ -552,27 +568,44 @@ class ConfigurationParser(object):
         option('--config', metavar='<path>', default=None,
                help='use given config file and disable config discovery')
 
+        check_group = OptionGroup(
+            parser,
+            'Error Check Options',
+            'Only one of --select, --ignore or --convention can be '
+            'specified. If none is specified, defaults to '
+            '`--convention=pep257`. These three options select the "basic '
+            'list" of error codes to check. If you wish to change that list '
+            '(for example, if you selected a known convention but wish to '
+            'ignore a specific error from it or add a new one) you can '
+            'use `--add-[ignore/select]` in order to do so.')
+        add_check = check_group.add_option
+
         # Error check options
-        option('--select', metavar='<codes>', default=None,
-               help='choose the basic list of checked errors by '
-                    'specifying which errors to check for (with a list of '
-                    'comma-separated error codes or prefixes). '
-                    'for example: --select=D101,D2')
-        option('--ignore', metavar='<codes>', default=None,
-               help='choose the basic list of checked errors by '
-                    'specifying which errors to ignore (with a list of '
-                    'comma-separated error codes or prefixes). '
-                    'for example: --ignore=D101,D2')
-        option('--convention', metavar='<name>', default=None,
-               help='choose the basic list of checked errors by specifying an '
-                    'existing convention. Possible conventions: {}'
-                    .format(', '.join(conventions)))
-        option('--add-select', metavar='<codes>', default=None,
-               help='amend the list of errors to check for by specifying '
-                    'more error codes to check.')
-        option('--add-ignore', metavar='<codes>', default=None,
-               help='amend the list of errors to check for by specifying '
-                    'more error codes to ignore.')
+        add_check('--select', metavar='<codes>', default=None,
+                  help='choose the basic list of checked errors by '
+                       'specifying which errors to check for (with a list of '
+                       'comma-separated error codes or prefixes). '
+                       'for example: --select=D101,D2')
+        add_check('--ignore', metavar='<codes>', default=None,
+                  help='choose the basic list of checked errors by '
+                       'specifying which errors to ignore out of all of the '
+                       'available error codes (with a list of '
+                       'comma-separated error codes or prefixes). '
+                       'for example: --ignore=D101,D2')
+        add_check('--convention', metavar='<name>', default=None,
+                  help='choose the basic list of checked errors by specifying '
+                       'an existing convention. Possible conventions: {}.'
+                       .format(', '.join(conventions)))
+        add_check('--add-select', metavar='<codes>', default=None,
+                  help='add extra error codes to check to the basic list of '
+                       'errors previously set by --select, --ignore or '
+                       '--convention.')
+        add_check('--add-ignore', metavar='<codes>', default=None,
+                  help='ignore extra error codes by removing them from the '
+                       'basic list previously set by --select, --ignore '
+                       'or --convention.')
+
+        parser.add_option_group(check_group)
 
         # Match clauses
         option('--match', metavar='<pattern>', default=None,