+++ src/tests/test_cases/test.py
@@ -27,6 +27,10 @@ class class_:
         pass
 
     @expect('D102: Missing docstring in public method')
+    def __new__(self=None):
+        pass
+
+    @expect('D107: Missing docstring in __init__')
     def __init__(self=None):
         pass