+++ docs/release_notes.rst
@@ -4,15 +4,40 @@ Release Notes
 **pydocstyle** version numbers follow the
 `Semantic Versioning <http://semver.org/>`_ specification.
 
-
-Current Development Version
+Current development version
 ---------------------------
+
+New features
+
+* Violations are now reported on the line where the docstring starts, not the
+  line of the ``def``/``class`` it corresponds to (#238, #83).
+
+
+2.1.1 - October 9th, 2017
+-------------------------
+
+Bug Fixes
+
+* Changed wheel configuration to be NOT universal, as #281 added
+  ``configparser`` as a dependency for Python 2.7.
+* Updated usage documentation.
+
+
+2.1.0 - October 8th, 2017
+-------------------------
 
 New Features
 
 * Public nested classes missing a docstring are now reported as D106 instead
   of D101 (#198, #261).
+* ``__init__`` methods missing a docstring are now reported as D107 instead of
+  D102 (#273, #277).
 * Added support for Python 3.6 (#270).
+* Specifying an invalid error code prefix (e.g., ``--select=D9``) will print
+  a warning message to ``stderr`` (#253, #279).
+* Configuration files now support multiple-lined entries (#250, #281).
+* Improved description of how error selection works in the help section
+  (#231, #283).
 
 Bug Fixes
 
@@ -21,6 +46,9 @@ Bug Fixes
 * Fixed an issue where if a first word in a docstring had Unicode characters
   and the docstring was not a unicode string, an exception would be raised
   (#258, #264).
+* Configuration files that were specified by CLI and don't contain a valid
+  section name will now issue a warning to ``stderr`` (#276, #280).
+* Removed D107 from the numpy convention (#288).
 
 
 2.0.0 - April 18th, 2017